requirements.txt
# Core STT
faster-whisper>=1.1.0
sounddevice>=0.5.1
numpy>=1.26

# VAD
sherpa-onnx>=1.12

# CUDA runtime (якщо CUDA Toolkit не встановлений)
nvidia-cublas-cu12
nvidia-cudnn-cu12

# LLM post-processing
requests>=2.31

# Hotkeys + tray
keyboard>=0.13
pystray>=0.19

# Text injection
pyperclip>=1.9
pyautogui>=0.9.54

# Notifications
win11toast>=0.36
pywin32>=306

# Process management
psutil>=5.9

# UI + config
customtkinter>=5.2
Pillow>=10.0
PyYAML>=6.0
